feat: refactor movie handler and add search functionality for free and YTS sources
