+++ api/library-service/src/handlers/health.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func HealthCheckHandler(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{"status": "ok", "service": "library-service"})
+}
